Address PR feedback: single RPC call, remove cache, use ExtensionRegistryInfo, remove redundant ide_id
