+++ src/services/telemetry/TelemetryService.ts
@@ -4,7 +4,6 @@ import { ShowMessageType } from "@shared/proto/host/window"
 import type { TaskFeedbackType } from "@shared/WebviewMessage"
 import * as os from "os"
 import { ClineAccountUserInfo } from "@/services/auth/AuthService"
-import { getIdeId } from "@/services/EnvService"
 import { Setting } from "@/shared/proto/index.host"
 import { Mode } from "@/shared/storage/types"
 import { version as extensionVersion } from "../../../package.json"
@@ -51,8 +50,6 @@ export type TelemetryMetadata = {
 	extension_version: string
 	/** The name of the host IDE or environment e.g. VSCode */
 	platform: string
-	/** IDE identifier string */
-	ide_id: string | undefined
 	/** The version of the host environment */
 	platform_version: string
 	/** The operating system type, e.g. darwin, win32. This is the value returned by os.platform() */
@@ -209,11 +206,9 @@ export class TelemetryService {
 			type: "posthog",
 		})
 		const hostVersion = await HostProvider.env.getHostVersion({})
-		const ideId = await getIdeId()
 		const metadata: TelemetryMetadata = {
 			extension_version: extensionVersion,
 			platform: hostVersion.platform || "unknown",
-			ide_id: ideId,
 			platform_version: hostVersion.version || "unknown",
 			os_type: os.platform(),
 			os_version: os.version(),
